extract test for rich text and linked items element to seprrate test suites
