--- src/__tests__/circularReferenceItemsResponse.json
@@ -1,62 +1,0 @@
-{
-  "items": [
-    {
-      "system": {
-        "id": "1a3aa6bd-7b4b-486f-86ab-7bfe359ad614",
-        "name": "Main project",
-        "codename": "main_project",
-        "language": "default",
-        "type": "project",
-        "sitemap_locations": [],
-        "last_modified": "2019-09-27T09:27:00.0135853Z"
-      },
-      "elements": {
-        "content": {
-          "type": "text",
-          "name": "Content",
-          "value": "Main project content"
-        },
-        "related_projects": {
-          "type": "modular_content",
-          "name": "Related projects",
-          "value": [
-            "main_project"
-          ]
-        }
-      }
-    }
-  ],
-  "modular_content": {
-    "main_project": {
-      "system": {
-        "id": "1a3aa6bd-7b4b-486f-86ab-7bfe359ad614",
-        "name": "Main project",
-        "codename": "main_project",
-        "language": "default",
-        "type": "project",
-        "sitemap_locations": [],
-        "last_modified": "2019-09-27T09:27:00.0135853Z"
-      },
-      "elements": {
-        "content": {
-          "type": "text",
-          "name": "Content",
-          "value": "Main project content"
-        },
-        "related_projects": {
-          "type": "modular_content",
-          "name": "Related projects",
-          "value": [
-            "main_project"
-          ]
-        }
-      }
-    }
-  },
-  "pagination": {
-    "skip": 0,
-    "limit": 0,
-    "count": 1,
-    "next_page": ""
-  }
-}
