--- src/__tests__/fakeTypesResponseWithRichTextElement.json
@@ -1,63 +1,0 @@
-{
-  "types": [
-    {
-      "system": {
-        "id": "b300f06d-d1e7-40b4-9dfc-e065e0455593",
-        "name": "Landing page",
-        "codename": "landing_page",
-        "last_modified": "2018-12-31T12:40:05.3545328Z"
-      },
-      "elements": {
-        "url_slug": {
-          "type": "url_slug",
-          "name": "URL slug"
-        },
-        "content": {
-          "type": "rich_text",
-          "name": "Content"
-        },
-        "landing_page_name": {
-          "type": "text",
-          "name": "Landing page name"
-        }
-      }
-    },
-    {
-      "system": {
-        "id": "45e6379c-858c-49aa-b293-b1218581931b",
-        "name": "Landing page image section",
-        "codename": "landing_page_image_section",
-        "last_modified": "2018-12-31T11:37:53.7435133Z"
-      },
-      "elements": {
-        "title": {
-          "type": "text",
-          "name": "Title"
-        },
-        "images": {
-          "type": "asset",
-          "name": "Images"
-        }
-      }
-    },
-    {
-      "system": {
-        "id": "ec706130-b0dd-401c-ae1a-10f2eeb2eb89",
-        "name": "Project",
-        "codename": "project",
-        "last_modified": "2018-12-28T12:29:09.4933517Z"
-      },
-      "elements": {
-        "related_projects": {
-          "type": "modular_content",
-          "name": "Related projects"
-        },
-        "content": {
-          "type": "text",
-          "name": "Content"
-        }
-      }
-    }
-  ],
-  "pagination": {}
-}
