+++ src/__tests__/richtext-resolution/richtextCircularReferenceFakeItemsResponse.json
@@ -1,0 +1,76 @@
+{
+  "items": [
+    {
+      "system": {
+        "id": "1a3aa6bd-7b4b-486f-86ab-7bfe359ad614",
+        "name": "Main project",
+        "codename": "main_project",
+        "language": "default",
+        "type": "project",
+        "sitemap_locations": [],
+        "last_modified": "2018-12-28T12:53:59.0753983Z"
+      },
+      "elements": {
+        "content": {
+          "type": "text",
+          "name": "Content",
+          "value": "Main project content"
+        }
+      }
+    },
+    {
+      "system": {
+        "id": "45267fb2-2e53-41de-a51f-77a28f0a05e5",
+        "name": "Simple Landing page",
+        "codename": "simple_landing_page",
+        "language": "default",
+        "type": "landing_page",
+        "sitemap_locations": [],
+        "last_modified": "2019-10-09T11:20:40.7244458Z"
+      },
+      "elements": {
+        "content": {
+          "type": "rich_text",
+          "name": "Content",
+          "images": {},
+          "links": {},
+          "modular_content": [
+            "simple_landing_page"
+          ],
+          "value": "<h1>Landing page</h1>\n<p><br>\nLorem ipsum dolor sit amet, consectetuer adipiscing elit. Duis condimentum augue id magna semper rutrum. Phasellus faucibus molestie nisl. Aliquam erat volutpat. Fusce consectetuer risus a nunc. Duis viverra diam non justo. Etiam commodo dui eget wisi. In enim a arcu imperdiet malesuada. Curabitur ligula sapien, pulvinar a vestibulum quis, facilisis vel sapien. Vivamus ac leo pretium faucibus. Nulla non arcu lacinia neque faucibus fringilla. Duis pulvinar. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos hymenaeos. Nulla est. Proin in tellus sit amet nibh dignissim sagittis. Aliquam erat volutpat. Donec iaculis gravida nulla. Integer malesuada. Proin mattis lacinia justo.</p>\n<object type=\"application/kenticocloud\" data-type=\"item\" data-rel=\"link\" data-codename=\"simple_landing_page\"></object>\n<p><br></p>"
+        }
+      }
+    }
+  ],
+  "modular_content": {
+    "simple_landing_page": {
+      "system": {
+        "id": "45267fb2-2e53-41de-a51f-77a28f0a05e5",
+        "name": "Simple Landing page",
+        "codename": "simple_landing_page",
+        "language": "default",
+        "type": "landing_page",
+        "sitemap_locations": [],
+        "last_modified": "2019-10-09T11:20:40.7244458Z"
+      },
+      "elements": {
+        "content": {
+          "type": "rich_text",
+          "name": "Content",
+          "images": {},
+          "links": {},
+          "modular_content": [
+            "simple_landing_page"
+          ],
+          "value": "<h1>Landing page</h1>\n<p><br>\nLorem ipsum dolor sit amet, consectetuer adipiscing elit. Duis condimentum augue id magna semper rutrum. Phasellus faucibus molestie nisl. Aliquam erat volutpat. Fusce consectetuer risus a nunc. Duis viverra diam non justo. Etiam commodo dui eget wisi. In enim a arcu imperdiet malesuada. Curabitur ligula sapien, pulvinar a vestibulum quis, facilisis vel sapien. Vivamus ac leo pretium faucibus. Nulla non arcu lacinia neque faucibus fringilla. Duis pulvinar. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos hymenaeos. Nulla est. Proin in tellus sit amet nibh dignissim sagittis. Aliquam erat volutpat. Donec iaculis gravida nulla. Integer malesuada. Proin mattis lacinia justo.</p>\n<object type=\"application/kenticocloud\" data-type=\"item\" data-rel=\"link\" data-codename=\"simple_landing_page\"></object>\n<p><br></p>"
+        }
+      }
+    }
+  },
+  "pagination": {
+    "skip": 0,
+    "limit": 0,
+    "count": 2,
+    "next_page": ""
+  }
+}
