+++ src/__tests__/richtext-resolution/richtextFakeItemsResponse.json
@@ -1,0 +1,137 @@
+{
+  "items": [
+    {
+      "system": {
+        "id": "1a3aa6bd-7b4b-486f-86ab-7bfe359ad614",
+        "name": "Main project",
+        "codename": "main_project",
+        "language": "default",
+        "type": "project",
+        "sitemap_locations": [],
+        "last_modified": "2018-12-28T12:53:59.0753983Z"
+      },
+      "elements": {
+        "content": {
+          "type": "text",
+          "name": "Content",
+          "value": "Main project content"
+        }
+      }
+    },
+    {
+      "system": {
+        "id": "45267fb2-2e53-41de-a51f-77a28f0a05e5",
+        "name": "Simple Landing page",
+        "codename": "simple_landing_page",
+        "language": "default",
+        "type": "landing_page",
+        "sitemap_locations": [],
+        "last_modified": "2019-10-09T11:29:41.3550672Z"
+      },
+      "elements": {
+        "content": {
+          "type": "rich_text",
+          "name": "Content",
+          "images": {
+            "98bf37b8-8be3-45e9-b284-293e205bf9a3": {
+              "image_id": "98bf37b8-8be3-45e9-b284-293e205bf9a3",
+              "description": null,
+              "url": "https://assets-eu-01.kc-usercontent.com:443/71be1cca-0be3-0159-fdfe-6cf11b092e68/6c7c8367-f38c-41a1-a41d-91a71dcc073a/Back%203_4%20right.jpg",
+              "width": 3010,
+              "height": 1700
+            }
+          },
+          "links": {
+            "1a3aa6bd-7b4b-486f-86ab-7bfe359ad614": {
+              "codename": "main_project",
+              "type": "project",
+              "url_slug": ""
+            }
+          },
+          "modular_content": [
+            "main_project",
+            "n26ddfb46_c2de_010e_3ad9_ec9aa8955144"
+          ],
+          "value": "<h1>Landing page</h1>\n<p><br>\n<strong>Lorem </strong><em>ipsum </em><sup>dolor </sup><sub>sit </sub>amet, consectetuer adipiscing elit. Duis condimentum augue id magna semper rutrum. Phasellus faucibus molestie nisl.&nbsp;</p>\n<ol>\n  <li>Aliquam erat volutpat.&nbsp;</li>\n</ol>\n<p><br></p>\n<p><a href=\"https://kontent.ai\">Link to web.</a></p>\n<p><br></p>\n<p><a data-asset-id=\"98bf37b8-8be3-45e9-b284-293e205bf9a3\" href=\"https://assets-eu-01.kc-usercontent.com:443/71be1cca-0be3-0159-fdfe-6cf11b092e68/6c7c8367-f38c-41a1-a41d-91a71dcc073a/Back%203_4%20right.jpg\">Link to Image</a></p>\n<p><br></p>\n<figure data-asset-id=\"98bf37b8-8be3-45e9-b284-293e205bf9a3\" data-image-id=\"98bf37b8-8be3-45e9-b284-293e205bf9a3\"><img src=\"https://assets-eu-01.kc-usercontent.com:443/71be1cca-0be3-0159-fdfe-6cf11b092e68/6c7c8367-f38c-41a1-a41d-91a71dcc073a/Back%203_4%20right.jpg\" data-asset-id=\"98bf37b8-8be3-45e9-b284-293e205bf9a3\" data-image-id=\"98bf37b8-8be3-45e9-b284-293e205bf9a3\" alt=\"\"></figure>\n<p>Inline component</p>\n<object type=\"application/kenticocloud\" data-type=\"item\" data-rel=\"component\" data-codename=\"n26ddfb46_c2de_010e_3ad9_ec9aa8955144\"></object>\n<p><br></p>\n<p>Projects (Link to content item):</p>\n<ul>\n  <li><a data-item-id=\"1a3aa6bd-7b4b-486f-86ab-7bfe359ad614\" href=\"\">Main project</a></li>\n</ul>\n<p>Aliquam ante. Nam quis nulla. Sed ac dolor sit amet purus malesuada congue. Curabitur vitae diam non enim vestibulum interdum. Phasellus et lorem id felis nonummy placerat. Etiam posuere lacus quis dolor.&nbsp;</p>\n<p><strong>Project (Link to existing content item)</strong></p>\n<object type=\"application/kenticocloud\" data-type=\"item\" data-rel=\"link\" data-codename=\"main_project\"></object>"
+        }
+      }
+    }
+  ],
+  "modular_content": {
+    "main_project": {
+      "system": {
+        "id": "1a3aa6bd-7b4b-486f-86ab-7bfe359ad614",
+        "name": "Main project",
+        "codename": "main_project",
+        "language": "default",
+        "type": "project",
+        "sitemap_locations": [],
+        "last_modified": "2018-12-28T12:53:59.0753983Z"
+      },
+      "elements": {
+        "content": {
+          "type": "text",
+          "name": "Content",
+          "value": "Main project content"
+        }
+      }
+    },
+    "n26ddfb46_c2de_010e_3ad9_ec9aa8955144": {
+      "system": {
+        "id": "26ddfb46-c2de-010e-3ad9-ec9aa8955144",
+        "name": "26ddfb46-c2de-010e-3ad9-ec9aa8955144",
+        "codename": "n26ddfb46_c2de_010e_3ad9_ec9aa8955144",
+        "language": "default",
+        "type": "landing_page_image_section",
+        "sitemap_locations": [],
+        "last_modified": "2019-10-09T11:29:41.3550672Z"
+      },
+      "elements": {
+        "title": {
+          "type": "text",
+          "name": "Title",
+          "value": "Cap"
+        },
+        "images": {
+          "type": "asset",
+          "name": "Images",
+          "value": [
+            {
+              "name": "Back 3_4 right.jpg",
+              "description": null,
+              "type": "image/jpeg",
+              "size": 1639188,
+              "url": "https://assets-eu-01.kc-usercontent.com:443/71be1cca-0be3-0159-fdfe-6cf11b092e68/6c7c8367-f38c-41a1-a41d-91a71dcc073a/Back%203_4%20right.jpg",
+              "width": 3010,
+              "height": 1700
+            },
+            {
+              "name": "Front.jpg",
+              "description": null,
+              "type": "image/jpeg",
+              "size": 1552869,
+              "url": "https://assets-eu-01.kc-usercontent.com:443/71be1cca-0be3-0159-fdfe-6cf11b092e68/1b6e8055-2952-4236-92d3-33e7774fb7d0/Front.jpg",
+              "width": 2365,
+              "height": 1826
+            },
+            {
+              "name": "Back 3_4.jpg",
+              "description": null,
+              "type": "image/jpeg",
+              "size": 1653072,
+              "url": "https://assets-eu-01.kc-usercontent.com:443/71be1cca-0be3-0159-fdfe-6cf11b092e68/feb9cb93-0d5e-4a22-879f-f6cb7dd6b67b/Back%203_4.jpg",
+              "width": 3010,
+              "height": 1700
+            }
+          ]
+        }
+      }
+    }
+  },
+  "pagination": {
+    "skip": 0,
+    "limit": 0,
+    "count": 2,
+    "next_page": ""
+  }
+}
